[YARPC Thrift Connector] An API module for YARPC connector with an OSCAR generated client
Summary:
We need to generate an OSCAR Thrift client in-order to talk to downstream services.
So we will create a presto-yarpc-connector-api maven module for Uber Presto project that will generate and compile a thrift client based on an presto_thrift_connector.thrift IDL.
Giving that it introduces dependency on Thrift complier, we will exclude presto-yarpc-connector-api from the parent module.

Ref T1739829

Reviewers: lu.niu

Reviewed By: lu.niu

Subscribers: zstewart, jenkins

Maniphest Tasks: T1739829

Differential Revision: https://code.uberinternal.com/D1832463