change next uri base to host name of coordinator
Summary:
change next uri base to host name of coordinator.

support feature as if submission request uri is localhost:5436 (muttley uri),  we can set the nexturi in queryResult to be 'prestomaster02-dca1:8080'

drawback is any hostname must be available to the client. This statement is true to all the clients (janus, piper, queryrunner, presto-cli). If any firewall is built in front of presto, this PR will cause problem.

Test Plan: test locally and it works

Reviewers: luoz, curt, bhasudha

Reviewed By: luoz

Tags: #presto

Differential Revision: https://code.uberinternal.com/D2001473