Restrict the metadata query search space
Summary:
Currently the BI tools ask for metadata of a single table through
generic JDBC API [1], but the input schema and table is always
considered as a pattern and we end up getting the metadata of all
Hive tables. This is a huge cost. One optimization we could do in
JDBC is when the schemaName or tableName is not a search pattern
make the filter as `=`. The HiveMetadata implementation already has
an optimization to restrict the search space based on the `=`
filter on either/both schemaName and tableName.

One other solution considered is specific to HiveMetadata where
make it understand the filters like `LIKE`, but this will be
Hive specific and will have to implement the same solution for
connectors in future.

[1] https://docs.oracle.com/javase/8/docs/api/java/sql/DatabaseMetaData.html#getColumns-java.lang.String-java.lang.String-java.lang.String-java.lang.String-

Reviewers: venkik

Reviewed By: venkik

Subscribers: jenkins

Maniphest Tasks: T2336945

Differential Revision: https://code.uberinternal.com/D2672167