Add decryption functionality to presto
Summary:
Part of the requirement of column access control is that Presto can have the decryption functionality to read the encrypted columns in Parquet. This change is to add decryption functionality to presto. This is the internal change for ticket: https://github.com/prestodb/presto/issues/12048

The original change in Parquet for encryption/decryption can be found https://code.uberinternal.com/rDAPARPE36f723212376257ce726a181a5417e99f8e72a44

The encryption specification for Parquet can be found here https://github.com/apache/parquet-format/blob/encryption/Encryption.md

Update:
Make the change to override/add instead of change the code directly, and add a switch(default false) to make this feature as opt-in one. Once the decryption code is well baked, we make another change to replace it completely with new parquet code.

Test Plan:
Local testing
Unit testing
Staging test will be done once it is landed because currently, the release-217 branch is missing features to run in staging.

Reviewers: luoz, lu.niu, baliga, venkik

Reviewed By: lu.niu

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2669729