Enable RG level run time and queuetime metrics on a periodic snapshot basis
Summary:
Details:
Current queuedtime metrics is updated only after a query finishes. in case of an outage queries can take as long as an hour to finish (either fail or succeed after queuing). Since this metrics comes after fact, its not useful to monitor the cluster when the queries are actually queuing. The goal of this task is to expose metrics when queueing is happening in the cluster.

Test Plan: unit tests included

Reviewers: venkik, lu.niu, curt, luoz

Reviewed By: venkik

Subscribers: jenkins, baliga

Maniphest Tasks: T2414791

Differential Revision: https://code.uberinternal.com/D2690655