Respond to the entire query when a header flag is set
Summary:
If X-Presto-Wait-For-Entire-Response-Ms is set then either the query
returns completely in resonse to the POST, or it times out and is killed.

If this header is absent, then the semantics are unchanged: Post returns
with query is created (But not yet submitted). The submission happens on
the first get.

I have tested this with both curl (and the header) and with the presto
client. It works just fine with the presto-client. That is we haven't
broken any client semantics here. Existing presto clients (both the
presto java one and the python ones) see no change from their end.

This turns out to be a bigger change: The problem is that the QueryState
can be done even when the QueryInfo isn't final yet because some task is
still running.

There is a race: a task can be running, the stage could be running, and
yet the root stage is done and therefore the query state is done.

This isn't a problem when the client keeps polling: the client will keep
polling as long as there is a next uri, which will always be send until
the final query info is reached.

So really the true indication of 'doneness' is final query info.

So I had to expose the final query info change also as a listenable
future (in the same way as the state change) and plumb that all the way
through from the Query, through the different QueryExecution
implementations.

Test Plan: In progress

Reviewers: venkik, lu.niu, curt

Reviewed By: curt

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2737957