[Clean up later] HACK to enforce single source partitioning for aggregation to remove redundant exchanges
Summary:
When these configs are set in the config.properties and the
connector, we will force the use single stage/single task when the scan
pipeline is present.

This should be done in a proper way using some optimizer rules perhaps,
(but I am not sure how to take in the pinot thing then). Perhaps the
scan pipeline needs to opine about the distribution strategy such that
the connector can create the right partitioning handle.

But I think atleast this diff is in a do no harm mode and perhaps safe to
go in with this setting off.

Definitely something we should fix later on, but for now perhaps it can
go into mainline ?

Test Plan: Unit tests and will test turning these flags on with neutrino

Reviewers: venkik, vivekb, lu.niu

Reviewed By: venkik

Subscribers: jenkins

Maniphest Tasks: T2882267

Differential Revision: https://code.uberinternal.com/D2796303