[Pinot bugfixes] Fix querying on laptop by talking to the rest proxy server instead of the broker/controller
Summary:
Cannot query using the controller because it does not allow querying.
And querying using the broker requires setting up the many ZK tunnels. So we
have to resurrect the rest proxy for local mode.

Similarly the rest proxy is needed to talk to the cross region hosts in phx.

Not enabling it for prod.

Test Plan: Test running neutrino on laptop

Reviewers: venkik, curt, lu.niu

Reviewed By: venkik

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2796205