[Pinot plugin] More pinot plugin bugfixes for broker page source (neutrino) mostly
Summary:
The big change/bugfix here is to change the way we create the
"column indices" for pinot. The old code was buggy and failed for tables
with mixed case. It was my mistake since I wrote it in the first place
:).

The smaller change here is to not allow neutrino queries (ie when
forSinglePageSource is set) to return more than max limit.

Ie when the query cannot be pushed down, don't drudge the data into
neutrino since it can't really handle that. This does not kick in for
non neutrino code paths

In the process of testing these changes, I had to fix some of the test
util code too.

Test Plan:
Added unit tests that stress the mixed case, limit and scan
only use cases

Reviewers: venkik

Reviewed By: venkik

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2816309