[Conditional local mode RPC] Bypass http calls if the node id is local
Summary:
If a global config is set and the node-uri matches the current
node, then use the task manager directly for fetching exchange data and
doing task calls. That is bypass the taskResource if these conditions
are satisfied.

In addition I have changed the backoff stuff such that the local mode
never undergoes a backoff ... Its still a bit hacky but does the job.

Can be enabled (default is disabled) by setting
task.bypass-http-for-local=true

Test Plan: Added a very basic unit test

Reviewers: venkik

Reviewed By: venkik

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2808039