[Pushdown testing] Adding SQL pushdown tests
Summary:
Basically, we had to expose more info about the "test tables" to the
metadata. It turns out that the TestingMetadata already had some of the pieces
and so I mostly resorted to that. Some new things had to be added but adding
them to the TestingMetadata broke other tests, so I introduced another base
class.

There is one shortcoming, for full disclosure: Now the schema was more strongly
matched and so some tests had to be tweaked to satisfy that.

This allows us to easily add more tests in the future. I find this interface
easier than crafting plan trees by hand.

(I think one of the newly added limit pushdown test depends on D2825351)

Test Plan: Added more unit tests

Reviewers: venkik

Reviewed By: venkik

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2834115