[Query session logging] Instrumentation framework to log key points in a query execution's lifecycle
Summary:
The basic usage is: session.getSessionLogger().log(() ->
String.format("some message")). These messages can be logged at the connector
(using the ConnectorSession) or at the coordinator or tasks (that have access
to the Session). The log messages are shipped as a part of the regular
QueryInfo from all the tasks and show up in the final json: there is one
"sessionLogEntries" for the QueryInfo and one per each of the tasks in the
query.

This is the more proper form of the hacky instrumentation introduced by the
name of HackTimerLog in release-216-agg branch. This works with both remote
tasks and local tasks (ie both when bypass-http is set to true or false).

This is disabled by default and can be enabled by using the header
X-Presto-Query-Logging-Size on the query.

I will follow up with a different diff that actually makes use of this.

Test Plan: Testing with local neutrino with the bypass-http set to true and false.

Reviewers: venkik, curt

Reviewed By: venkik

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2848161