Revert "[Limit Pushdown] Split the limit into a partial and the final so that the partial one can be pushed down"
Summary:
I would like to redo the way I am handling limit pushdown.

This diff was introduced as a way to fix the limit pushdown issue, but
it broke the project pushdown: Project pushdown now isn't getting
triggered because there is a final limit in the picture.

So I will revert this and handle this at the connector level directly.

This reverts commit d39457b962c140618483d5a5314ba079705f5535.

Test Plan: Unit tests

Reviewers: venkik

Reviewed By: venkik

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2871837