[Pinot connector] Use scan parallelism finder to decide whether to push a final limit down into pinot
Summary:
The bulk of the code is refactoring needed to prevent
duplication during production and (new) unit tests.

I am opting to use the nice ScanParallelism finder instead of simple
checks like hasAggregation etc so that we have less fixing to do when we
add support for sort/order etc.

The logic for deciding whether to add the final limit has been promoted
to PinotMetadata (that has the full config etc) instead of the
PinotQueryGenerator. The upside of this is that this checking only
happens during the optimizer and that the PQL is printed with the limit
during the explain plan.

With this both the neutrino (scan Parallelism disabled) and photon use
cases (pushing partial limit introduced by exchanges down) are handled.

Test Plan:
Added unit tests for this and tested in neutrino/photon
codepaths.

Reviewers: venkik

Reviewed By: venkik

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2871851