[Pinot] Separate expression converter into filter and project expression converter
Summary:
Currently we use a single expression converter which causes issues as the not all
expressions are supported in both project and filter. For example IN <list> is
supported only in filter expressions and not in project expressions. Similarly
arithmetic expressions are only supported in project and not in filter.

Reviewers: dagrawal

Reviewed By: dagrawal

Subscribers: jenkins

Maniphest Tasks: T3111919

Differential Revision: https://code.uberinternal.com/D2925185