[Pinot] NOT operator pushdown is supported only on top of IN operator
Summary:
For query:
`SELECT accountUUID ,clientUUID from pinotadhoc.eats_job_state.eats_job_state WHERE regionid NOT IN (574, 575)`

Presto separates the `col NOT IN (list)` into `NOT (col IN (list))` format. Currently in pushdown code we are
converting it as `NOT (col IN (list))` which is not supported by Pinot. Also NOT operator is not supported
anywhere else in Pinot expressions. Fix the handling of `NOT (col IN (list))` to rewrite as `col NOT IN (list)`

Reviewers: dagrawal

Reviewed By: dagrawal

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2926383