[Pinot] Introduce a different default limit for aggregation vs selects
Summary:
so it turns out that selection limits are free in pinot but
aggregation limits are costly since they require larger heap allocation.

Also we don't want to bound the non aggregation, non broker limit, so
the limit for that is now Int_max as in 203.

Test Plan: unit tests

Reviewers: venkik

Reviewed By: venkik

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2932059