[Pinot] Switch to using the (cached) controller http get instead of using ZK to access the brokers for a table
Summary:
Using the ZK directly is error prone and brittle. It will also
cause issues once the ZK servers are replaced since our configs would
have to updated. Remove this depandance on ZK and directly call the
published controller URI. Cache the broker list for a table for perf.

Once this is landed and deployed, we would have to remove the configs
zk-uri and pinot-cluster that are removed in this diff, otherwise the
server won't start.

Test Plan: Tested with pinotphxstg

Reviewers: venkik, haibow, lu.niu

Reviewed By: haibow, lu.niu

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2932203