[Plan caching] Disable plan caching when there are time functions present
Summary:
Because otherwise, we will get wrong answers when the time
changes and we are using the stale plan. Presto planner replaces the
current time with the literal session start time early on in the
planning process. (And that cannot be disabled because other optimizers
like tuple domain etc rely on that). This does not happen for non
deterministic functions (like random)

This isn't ideal :(, but atleast better than globally disabling plan
caching for neutrino.

Basically, for presto, time is constant (as determined by the session
start time) and hence can be replaced.

In any case this setting does no harm for Photon: Plan caching is
disabled globally.

I agree that it is hacky, but is a smaller change than introducing yet
another annotation to the SqlFunction/FunctionManager machinery which
would really snowball.

Also a bit of refactoring: For some reason, the SqlCachingPlanner was a
part of executions package, moved it to planner as it should be. Also
added unit test !

Test Plan: unit test and manual verification

Reviewers: burak, venkik, curt

Reviewed By: curt

Subscribers: curt, lu.niu, jenkins

Maniphest Tasks: T2923187

Differential Revision: https://code.uberinternal.com/D2955341