[Ares][Rta] Introduce max limit for Ares and use retention based default time filter if time filter is unspecified
Summary:
Sorry that these two diffs became a bit intertwined, and I got
tired of the merge conflicts when trying to untangle them. But there are
two Ares and RTA related changes here:

- Have a max limit for non aggregated ares queries. Ares would impose a
default max limit of 1000 in such a case, and we want that to be
configurable
- When the time filter is unspecified, just infer that based on the
current query start-time and the retention as contained in the metadata.
This prevents errors when a time filter is not specified.

Depends on D2991335

Test Plan: Added unit tests for all of the ares query and expression generation including the new stuff in this diff. And the unit tests actually helped find and fix bugs !

Reviewers: burak, venkik, lu.niu

Reviewed By: lu.niu

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2991341