[Pushdown] Pushing right side joins into the left's table scan pipeline
Summary:
A restrictive join pushdown framework that lets you push the right side scan pipeline into the left. I will follow this diff up with an Ares/RTA side join pushdown implementation. The default is that joins cannot be pushed down.

I considered an alternative version of this that allows the join to be flipped if the left table is the dimension table. But that didn't pan out, so just implementing a simpler version that fits the current use case of ares fact-dim join.

Test Plan: The subsequent diff will actually be testing this. Currently only tested manually with ares/rta

Reviewers: burak, venkik, vivekb

Reviewed By: burak

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D2991351